+++ src/main/vuejs/pages/staff/staff.vue
@@ -1,23 +1,12 @@
 <template>
-  <div class="staff">
-    <vheader></vheader>
-    <vslide></vslide>        
+  <div class="staff">        
     <div class="container">
       <p>人员页</p>
     </div>
-    <vfooter></vfooter>
   </div>
 </template>
 <script type="text/javascript">
-  import vslide from 'components/slide/Slide'
-  import vheader from 'components/header/header'
-  import vfooter from 'components/footer/footer'
   export default {
-    components: {
-      vslide,
-      vheader,
-      vfooter
-    }
   }
 </script>
 <style type="text/css">
